Cameras are now components
This is near totally untested, but cameras are now attachable to nodes
and should render the scene from wherever their attached object is.

I'm not totally happy with the current draw process, which is loop over
all the objects in the scene calling update functions with the parent
node's combined transform, then when a camera is hit going back to the
base node and going through it all again only with the camera's
transform and perspective tacked on to the matrix. It seems to work
though, if I think of a better system I'll change it.