First person controls on the camera are fixed
Turns out the camera was just doing the opposite of whatever I wanted.
So inverting the camera's transform and passing that into the draw call
along with the perspective matrix works like a charm.

Also, the operator to multiply a Matrix4 by a Vector3 has been replaced
with two function, one to rotate the vector by the matrix and the other
to translate it. That should make more logical sense.