Renamed the project
Until now this project has been named glTest2, in honour of the time I
accidently told the linker to use the debug libs for the release version
and vice versa and decided it would be easier to start again then fix
it.

Now it's properly named dsgl, although I'm sure I did this wrong.